+++ app/backend/tastify_backend/api_router.py
@@ -1,7 +1,7 @@
 from django.urls import include, path
 from rest_framework.routers import DefaultRouter
 from apps.menu.views import CategorieViewSet, PlatViewSet
-from apps.tables.views import TableViewSet
+from apps.tables.views import TableViewSet, PlanTextViewSet
 from apps.commandes.views import CommandeViewSet, CommandeLigneViewSet
 from apps.hr.views import EmployeViewSet, ShiftViewSet, OffreEmploiViewSet, CandidatureViewSet
 from apps.paiements.urls import urlpatterns as paiements_urlpatterns
@@ -14,6 +14,7 @@ router = DefaultRouter()
 router.register(r'categories', CategorieViewSet, basename='categorie')
 router.register(r'plats', PlatViewSet, basename='plat')
 router.register(r'tables', TableViewSet, basename='table')
+router.register(r'plan-texts', PlanTextViewSet, basename='plantext')
 router.register(r'commandes', CommandeViewSet, basename='commande')
 router.register(r'commandelignes', CommandeLigneViewSet, basename='commandeligne')
 router.register(r'employes', EmployeViewSet, basename='employe')